# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabeleleila Leila",
    "description": "Sistema de agendamento para o salão Cabeleleila Leila",
    "version": "1.0.0"
  },
  "paths": {
...
    "/client/edit-appointment/{appointment_id}": {
      "get": {
        "tags": [
          "Páginas"
        ],
        "summary": "Client Edit Appointment",
        "description": "Página para editar agendamento.",
        "operationId": "client_edit_appointment_client_edit_appointment__appointment_id__get",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Login",
        "description": "Autentica um usuário e retorna o token.",
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/google/login": {
      "get": {
        "tags": [
          "Autenticação Google"
        ],
        "summary": "Google Login",
        "description": "Inicia o fluxo OAuth do Google. \nPassa a role no 'state' para saber depois se é login de admin ou client.",
        "operationId": "google_login_api_auth_google_login_get",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "CLIENT",
              "title": "Role"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/google/callback": {
      "get": {
        "tags": [
          "Autenticação Google"
        ],
        "summary": "Google Callback",
        "description": "Recebe o callback do Google, cria/atualiza o usuário e faz o login interno.",
        "operationId": "google_callback_api_auth_google_callback_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/services": {
      "get": {
        "tags": [
          "Serviços"
        ],
        "summary": "List Services",
        "description": "Lista serviços disponíveis.",
        "operationId": "list_services_api_services_get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "only_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Only Active"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Busca por nome ou descrição",
              "title": "Search"
            },
            "description": "Busca por nome ou descrição"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Serviços"
        ],
        "summary": "Create Service",
        "description": "Cria um novo serviço (admin apenas).",
        "operationId": "create_service_api_services_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/services/{service_id}": {
      "get": {
        "tags": [
          "Serviços"
        ],
        "summary": "Get Service",
        "description": "Busca um serviço por ID.",
        "operationId": "get_service_api_services__service_id__get",
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Service Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Serviços"
        ],
        "summary": "Update Service",
        "description": "Atualiza um serviço (admin apenas).",
        "operationId": "update_service_api_services__service_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Service Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ServiceCreate": {
        "properties": {
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "duracao_minutos": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Duracao Minutos",
            "description": "Duração em minutos"
          },
          "preco": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Preco",
            "description": "Preço do serviço"
          }
        },
        "type": "object",
        "required": [
          "nome",
          "duracao_minutos",
          "preco"
        ],
        "title": "ServiceCreate",
        "description": "Schema para criação de serviço."
      },
      "ServiceResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "duracao_minutos": {
            "type": "integer",
            "title": "Duracao Minutos"
          },
          "preco": {
            "type": "number",
            "title": "Preco"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "duracao_minutos",
          "preco",
          "ativo"
        ],
        "title": "ServiceResponse",
        "description": "Schema de resposta de serviço."
      },
      "ServiceUpdate": {
        "properties": {
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "descricao": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descricao"
          },
          "duracao_minutos": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duracao Minutos"
          },
          "preco": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Preco"
          },
          "ativo": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ativo"
          }
        },
        "type": "object",
        "title": "ServiceUpdate",
        "description": "Schema para atualização de serviço."
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "TokenResponse",
        "description": "Schema de resposta de token JWT."
      },
...
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "senha": {
            "type": "string",
            "title": "Senha"
          }
        },
        "type": "object",
        "required": [
          "email",
          "senha"
        ],
        "title": "UserLogin",
        "description": "Schema para login."
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nome": {
            "type": "string",
            "title": "Nome"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "tipo": {
            "$ref": "#/components/schemas/UserType"
          },
          "ativo": {
            "type": "boolean",
            "title": "Ativo"
          },
          "criado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Criado Em"
          },
          "atualizado_em": {
            "type": "string",
            "format": "date-time",
            "title": "Atualizado Em"
          },
          "has_google_calendar": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Google Calendar",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "nome",
          "email",
          "tipo",
          "ativo",
          "criado_em",
          "atualizado_em"
        ],
        "title": "UserResponse",
        "description": "Schema de resposta de usuário (sem senha)."
      },
      "UserType": {
        "type": "string",
        "enum": [
          "CLIENT",
          "ADMIN"
        ],
        "title": "UserType",
        "description": "Tipos de usuário no sistema."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}